moved player movement code into its own class

+++ src/sprite_actor.cpp
@@ -41,7 +41,7 @@ void SpriteActor::Draw(sf::RenderWindow& window)
 sf::IntRect SpriteActor::GetRect()
 {
 	sf::IntRect rect = m_sprite.getTextureRect();
-	rect.left = m_sprite.getPosition().x;
-	rect.top = m_sprite.getPosition().y;
+	rect.left = (int) m_sprite.getPosition().x;
+	rect.top = (int) m_sprite.getPosition().y;
 	return rect;
 }
